fix avelin warnings

+++ common/LuaHandler.cpp
@@ -30,16 +30,17 @@ std::pair<uint16_t, uint16_t> LuaHandler::getTarget(std::pair<double, double> mo
 		return (std::make_pair(0, 0));
 	for (int i = 0; i < 10; i++)
 		playerPos[i] = -1;
-	playerPos[0] = mobPos.first;
-	playerPos[1] = mobPos.second;
+	playerPos[0] = static_cast<int>(mobPos.first);
+	playerPos[1] = static_cast<int>(mobPos.second);
 	c = 2;
 	for (auto & element : players) {
-		playerPos[c] = element.x;
-		playerPos[c + 1] = element.y;
+		playerPos[c] = static_cast<int>(element.x);
+		playerPos[c + 1] = static_cast<int>(element.y);
 		c += 2;
 	}
 	_lua->callFunction("getTarget", 10, playerPos, result);
-	if (result > players.size())
+	if (result > static_cast<int>(players.size()))
 		return (std::make_pair(0, 0));
-	return (std::make_pair(players[result].x, players[result].y));
+	return (std::make_pair(static_cast<uint16_t>(players[result].x),
+			static_cast<int>(players[result].y)));
 }
